Add vanilla CSS support to DropdownMenu (#526)
* Add / update demos

* Update examples
