+++ components/demos/index.jsx
@@ -7,6 +7,7 @@ export { default as CheckboxDemo } from './Checkbox/stitches';
 export { default as CollapsibleDemo } from './Collapsible/stitches';
 export { default as ContextMenuDemo } from './ContextMenu/stitches';
 export { default as DialogDemo } from './Dialog/stitches';
+export { default as DropdownMenuDemo } from './DropdownMenu/stitches';
 export { default as DropdownMenuDemo } from './DropdownMenu';
 export { default as HoverCardDemo } from './HoverCard';
 export { default as LabelDemo } from './Label';
